+++ install.sh
@@ -15,6 +15,7 @@ NO_EXEC_TEST="$4"
 NAME_PROJECT="$5"
 GEOSERVER_VERSION="2.19.6"
 
+
 rem_sig_hand() {
     trap "" ERR
     echo 'signal trapped'
@@ -29,11 +30,17 @@ if [ $GEM_SET_DEBUG ]; then
     set -x
 fi
 
+# virtual env
+python3.8 -m venv venv
+source ./venv/bin/activate
+
+# install django 3.2.12 
+pip install Django==3.2.12
+
 cd $HOME
 
-# sudo rm -rf oq-moon openquakeplatform geonode-project geoserver geoserver_data oq || true
 sudo rm -rf oq-moon openquakeplatform geonode-project geoserver oq || true
-sudo rm -rf oq-platform3/openquakeplatform/geonode || true
+sudo rm -rf oq-platform3/openquakeplatform_src/geonode || true
 # sudo rm oq-platform3/geoserver_data.tar.gz || true
 sudo rm -rf data || true
 sudo rm /usr/share/keyrings/docker-archive-keyring.gpg || true
@@ -41,16 +48,8 @@ sudo rm /usr/share/keyrings/docker-archive-keyring.gpg || true
 # display each command before executing it
 # set -x
 
-sudo apt-get -y update
-sudo apt-get -y upgrade
-
 #install git and ca-certificates
-sudo apt-get -y install git ca-certificates wget
-
-cd $GEM_GIT_PACKAGE
-cp .env.sample .env
-
-cd $HOME
+udo apt-get -y install git ca-certificates wget
 
 inst_docker () {
     # install requirements for docker
@@ -74,56 +73,55 @@ inst_docker
 #clone of repo 3.3.x 
 git clone -b 3.3.x https://github.com/GeoNode/geonode-project.git $HOME/geonode-project
 
-# uwsgi check if fixed in branch 3.3.3 of geonode-project
-cp $HOME/oq-platform3/uwsgi_files/create_envfile.py $HOME/geonode-project/
-cp $HOME/oq-platform3/uwsgi_files/geonode.conf.envsubst $HOME/geonode-project/docker/nginx/
-cp $HOME/oq-platform3/uwsgi_files/uwsgi.ini $HOME/geonode-project/src/
-
-cp $HOME/oq-platform3/.env $HOME/geonode-project/
-cp $HOME/oq-platform3/Dockerfile $HOME/geonode-project/
-cp $HOME/oq-platform3/docker-compose.yml $HOME/geonode-project/
-cp $HOME/oq-platform3/local_settings.py.tmpl $HOME/geonode-project/
-cp -pr $HOME/oq-platform3/pla_common $HOME/geonode-project/
-cp -pr $HOME/oq-platform3/data_commands $HOME/geonode-project/
-cp -pr $HOME/oq-platform3/openquakeplatform/ghec_viewer $HOME/geonode-project/
-cp -pr $HOME/oq-platform3/openquakeplatform/isc_viewer $HOME/geonode-project/
-
-# template
-mkdir $HOME/geonode-project/openquakeplatform
-mkdir $HOME/geonode-project/openquakeplatform/templates
-cp -pr $HOME/oq-platform3/openquakeplatform/templates/* $HOME/geonode-project/openquakeplatform/templates/
-cp $HOME/oq-platform3/openquakeplatform/urls.py $HOME/geonode-project/openquakeplatform/
-# cp -pr $HOME/oq-platform3/openquakeplatform/layers $HOME/geonode-project/openquakeplatform/
-# cp -pr $HOME/oq-platform3/openquakeplatform/maps $HOME/geonode-project/openquakeplatform/
-
-#static
-mkdir $HOME/geonode-project/openquakeplatform/static
-mkdir $HOME/geonode-project/openquakeplatform/static/css
-cp $HOME/oq-platform3/openquakeplatform/static/css/oqplatform.css $HOME/geonode-project/openquakeplatform/static/css/
-cp -pr $HOME/oq-platform3/openquakeplatform/static/geonode/img $HOME/geonode-project/openquakeplatform/static/
-
-
-# cp -pr $HOME/oq-platform3/gs_data/data $HOME/geonode-project/openquakeplatform/
-wget https://ftp.openquake.org/oq-platform3/data.tar.gz
-tar zxf data.tar.gz
-sudo cp -pr $HOME/oq-platform3/gs_data $HOME/geonode-project/openquakeplatform/
-
-rm data.tar.gz gs_data.tar.gz | true
-rm -rf data | true
-cp -pr $HOME/oq-platform3/openquakeplatform/bin $HOME/geonode-project
-cp -pr $HOME/oq-platform3/openquakeplatform/common $HOME/geonode-project
-
-# virtual env
-python3.8 -m venv $HOME/platform3
-source $HOME/platform3/bin/activate
-
-# install django
-pip install Django==3.2.12
-
-# start django project
-django-admin startproject -v 3 --template=$HOME/geonode-project -e py,sh,md,rst,json,yml,ini,env,sample,properties -n monitoring-cron -n Dockerfile $NAME_PROJECT
+if [ "$IS_STARTPROJECT" ]; then
+    # uwsgi check if fixed in branch 3.3.3 of geonode-project
+    cp $HOME/oq-platform3/uwsgi_files/create_envfile.py $HOME/geonode-project/
+    cp $HOME/oq-platform3/uwsgi_files/geonode.conf.envsubst $HOME/geonode-project/docker/nginx/
+    cp $HOME/oq-platform3/uwsgi_files/uwsgi.ini $HOME/geonode-project/src/
+    
+    cp $HOME/oq-platform3/.env $HOME/geonode-project/
+    cp $HOME/oq-platform3/Dockerfile $HOME/geonode-project/
+    cp $HOME/oq-platform3/docker-compose.yml $HOME/geonode-project/
+    cp $HOME/oq-platform3/local_settings.py.tmpl $HOME/geonode-project/
+    cp -pr $HOME/oq-platform3/pla_common $HOME/geonode-project/
+    cp -pr $HOME/oq-platform3/data_commands $HOME/geonode-project/
+    cp -pr $HOME/oq-platform3/openquakeplatform_src/ghec_viewer $HOME/geonode-project/
+    cp -pr $HOME/oq-platform3/openquakeplatform_src/isc_viewer $HOME/geonode-project/
+    
+    # template
+    mkdir -p $HOME/geonode-project/openquakeplatform/templates
+    cp -pr $HOME/oq-platform3/openquakeplatform_src/templates/* $HOME/geonode-project/openquakeplatform/templates/
+    cp $HOME/oq-platform3/openquakeplatform_src/urls.py $HOME/geonode-project/openquakeplatform/
+    
+    #static
+    mkdir -p $HOME/geonode-project/openquakeplatform/static/css
+    cp $HOME/oq-platform3/openquakeplatform_src/static/css/oqplatform.css $HOME/geonode-project/openquakeplatform/static/css/
+    cp -pr $HOME/oq-platform3/openquakeplatform_src/static/geonode/img $HOME/geonode-project/openquakeplatform/static/
+    
+    
+    wget https://ftp.openquake.org/oq-platform3/data.tar.gz
+    tar zxf data.tar.gz
+    sudo cp -pr $HOME/oq-platform3/gs_data $HOME/geonode-project/openquakeplatform/
+    
+    rm data.tar.gz gs_data.tar.gz | true
+    rm -rf data | true
+    cp -pr $HOME/oq-platform3/openquakeplatform_src/bin $HOME/geonode-project
+    cp -pr $HOME/oq-platform3/openquakeplatform_src/common $HOME/geonode-project
+    
+    cd $GEM_GIT_PACKAGE
+    cp .env.sample .env
+    
+    # start django project
+    django-admin startproject -v 3 --template=$HOME/geonode-project -e py,sh,md,rst,json,yml,ini,env,sample,properties -n monitoring-cron -n Dockerfile $NAME_PROJECT
+    exit 0
+fi    
 
 cd $NAME_PROJECT
+if [ ! -f .env ]; then
+    secret_key="$(python -c "from django.core.management.utils import get_random_secret_key; print(get_random_secret_key())")"
+    sed "s/{{secret_key}}/$secret_key/g" <.env.tmpl >.env
+fi    
+exit 123
 
 # Geoserver
 wget --no-check-certificate --progress=bar:force:noscroll https://artifacts.geonode.org/geoserver/${GEOSERVER_VERSION}/geoserver.war -O geoserver.war
